refactor(ai): modularize chat components and runner to resolve CodeScene complexity gates
