+++ frontend/src/routes/_layout/ai.tsx
@@ -1,38 +1,22 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
 import { createFileRoute } from "@tanstack/react-router"
-import {
-  Archive,
-  ArchiveRestore,
-  Check,
-  ChevronDown,
-  MoreHorizontal,
-  Pencil,
-  Search,
-  SquarePen,
-  Trash2,
-  X,
-} from "lucide-react"
 import * as React from "react"
 import { AiThreadsService, type ChatThreadPublic } from "@/client"
+import {
+  AIThreadsSidebar,
+  type SortOption,
+} from "@/components/Chat/AIThreadsSidebar"
 import { ChatMessage } from "@/components/Chat/ChatMessage"
+import { DeleteThreadConfirmDialog } from "@/components/Chat/DeleteThreadConfirmDialog"
 import { PromptForm } from "@/components/Chat/PromptForm"
 import { QuestionCard } from "@/components/Chat/QuestionCard"
+import { RenameThreadDialog } from "@/components/Chat/RenameThreadDialog"
 import { Suggestions } from "@/components/Chat/Suggestions"
 import type {
   AskUserAnswer,
   AskUserToolPart,
   ChatUIMessage,
 } from "@/components/Chat/types"
-import { Button } from "@/components/ui/button"
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog"
 import {
   MessageScroller,
   MessageScrollerButton,
@@ -42,7 +26,6 @@ import {
   MessageScrollerViewport,
 } from "@/components/ui/message-scroller"
 import { useAIChatStream } from "@/hooks/useAIChatStream"
-import { cn } from "@/lib/utils"
 
 export const Route = createFileRoute("/_layout/ai")({
   component: AIPage,
@@ -55,376 +38,48 @@ export const Route = createFileRoute("/_layout/ai")({
   }),
 })
 
-function DeleteThreadConfirmDialog({
-  thread,
-  isOpen,
-  isPending,
-  onClose,
-  onConfirm,
-}: {
-  thread: ChatThreadPublic | null
-  isOpen: boolean
-  isPending: boolean
-  onClose: () => void
-  onConfirm: () => void
-}) {
-  if (!thread) return null
+const AI_MODEL_STORAGE_KEY = "linkx_ai_selected_model"
 
-  return (
-    <Dialog open={isOpen} onOpenChange={(open) => !open && onClose()}>
-      <DialogContent className="sm:max-w-md">
-        <DialogHeader>
-          <DialogTitle>Delete Chat?</DialogTitle>
-          <DialogDescription>
-            This will permanently delete{" "}
-            <span className="font-semibold text-foreground">
-              "{thread.title}"
-            </span>{" "}
-            and its entire conversation history from your workspace. This action
-            cannot be undone.
-          </DialogDescription>
-        </DialogHeader>
-
-        <DialogFooter className="mt-4 gap-2.5 sm:gap-2.5">
-          <DialogClose asChild>
-            <Button
-              variant="outline"
-              size="sm"
-              disabled={isPending}
-              onClick={onClose}
-              className="cursor-pointer"
-            >
-              Cancel
-            </Button>
-          </DialogClose>
-          <Button
-            variant="destructive"
-            size="sm"
-            disabled={isPending}
-            onClick={onConfirm}
-            className="font-semibold cursor-pointer"
-          >
-            {isPending ? "Deleting…" : "Delete Chat"}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  )
-}
-
-function RenameThreadDialog({
-  thread,
-  isOpen,
-  isPending,
-  onClose,
-  onConfirm,
-}: {
-  thread: ChatThreadPublic | null
-  isOpen: boolean
-  isPending: boolean
-  onClose: () => void
-  onConfirm: (newTitle: string) => void
-}) {
-  const [titleInput, setTitleInput] = React.useState("")
-
-  React.useEffect(() => {
-    if (thread) {
-      setTitleInput(thread.title)
-    }
-  }, [thread])
-
-  if (!thread) return null
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    const trimmed = titleInput.trim()
-    if (trimmed) {
-      onConfirm(trimmed)
+function getInitialStoredModel(): string {
+  if (typeof window !== "undefined" && typeof localStorage !== "undefined") {
+    try {
+      const saved = localStorage.getItem(AI_MODEL_STORAGE_KEY)
+      if (saved) return saved
+    } catch {
+      // ignore
     }
   }
-
-  return (
-    <Dialog open={isOpen} onOpenChange={(open) => !open && onClose()}>
-      <DialogContent className="sm:max-w-md">
-        <form onSubmit={handleSubmit}>
-          <DialogHeader>
-            <DialogTitle>Rename Chat</DialogTitle>
-            <DialogDescription>
-              Enter a new title for this conversation.
-            </DialogDescription>
-          </DialogHeader>
-
-          <div className="py-4">
-            <input
-              type="text"
-              autoFocus
-              value={titleInput}
-              onChange={(e) => setTitleInput(e.target.value)}
-              placeholder="Chat title"
-              disabled={isPending}
-              className="w-full rounded-md border border-input bg-background px-3 py-2 text-sm text-foreground shadow-sm transition-colors placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring"
-            />
-          </div>
-
-          <DialogFooter className="gap-2.5 sm:gap-2.5">
-            <DialogClose asChild>
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                disabled={isPending}
-                onClick={onClose}
-                className="cursor-pointer"
-              >
-                Cancel
-              </Button>
-            </DialogClose>
-            <Button
-              type="submit"
-              size="sm"
-              disabled={isPending || !titleInput.trim()}
-              className="font-semibold cursor-pointer"
-            >
-              {isPending ? "Saving…" : "Save"}
-            </Button>
-          </DialogFooter>
-        </form>
-      </DialogContent>
-    </Dialog>
-  )
+  return "gemini-3.6-flash-high"
 }
 
-function ThreadActionsMenu({
-  isArchived,
-  onToggleArchive,
-  onStartRename,
-  onDelete,
-}: {
-  isArchived: boolean
-  onToggleArchive: () => void
-  onStartRename: () => void
-  onDelete: () => void
-}) {
-  return (
-    <div
-      role="menu"
-      tabIndex={-1}
-      className="absolute right-0 top-7 z-50 min-w-28 rounded-2xl border border-border bg-popover p-1 shadow-md animate-in fade-in-0 zoom-in-95"
-    >
-      <button
-        type="button"
-        role="menuitem"
-        onClick={(e) => {
-          e.stopPropagation()
-          onToggleArchive()
-        }}
-        className="flex w-full items-center gap-2 rounded-lg px-2.5 py-1.5 text-xs text-popover-foreground hover:bg-accent hover:text-accent-foreground cursor-pointer font-medium"
-      >
-        {isArchived ? (
-          <>
-            <ArchiveRestore className="size-3.5" />
-            <span>Unarchive</span>
-          </>
-        ) : (
-          <>
-            <Archive className="size-3.5" />
-            <span>Archive</span>
-          </>
-        )}
-      </button>
-      <button
-        type="button"
-        role="menuitem"
-        onClick={(e) => {
-          e.stopPropagation()
-          onStartRename()
-        }}
-        className="flex w-full items-center gap-2 rounded-lg px-2.5 py-1.5 text-xs text-popover-foreground hover:bg-accent hover:text-accent-foreground cursor-pointer font-medium"
-      >
-        <Pencil className="size-3.5" />
-        <span>Rename</span>
-      </button>
-      <button
-        type="button"
-        role="menuitem"
-        onClick={(e) => {
-          e.stopPropagation()
-          onDelete()
-        }}
-        className="flex w-full items-center gap-2 rounded-lg px-2.5 py-1.5 text-xs text-destructive hover:bg-destructive/10 cursor-pointer font-medium"
-      >
-        <Trash2 className="size-3.5" />
-        <span>Delete</span>
-      </button>
-    </div>
-  )
-}
-
-type SortOption = "recent" | "oldest" | "title" | "messages"
-
-function SortFilterMenu({
-  sortOrder,
-  onSelectSort,
-}: {
-  sortOrder: SortOption
-  onSelectSort: (order: SortOption) => void
-}) {
-  const options: { id: SortOption; label: string }[] = [
-    { id: "recent", label: "Most Recent" },
-    { id: "oldest", label: "Oldest First" },
-    { id: "title", label: "Alphabetical (A–Z)" },
-    { id: "messages", label: "Most Messages" },
-  ]
-
-  return (
-    <div
-      role="menu"
-      tabIndex={-1}
-      className="absolute right-0 top-8 z-50 min-w-44 rounded-2xl border border-border bg-popover p-1 shadow-md animate-in fade-in-0 zoom-in-95 text-xs"
-    >
-      <div className="px-2.5 py-1.5 font-semibold text-muted-foreground text-[11px] border-b border-border/40 select-none">
-        Sort by
-      </div>
-      <div className="p-0.5 space-y-0.5">
-        {options.map((opt) => (
-          <button
-            key={opt.id}
-            type="button"
-            role="menuitem"
-            onClick={(e) => {
-              e.stopPropagation()
-              onSelectSort(opt.id)
-            }}
-            className={cn(
-              "flex w-full items-center justify-between gap-2 rounded-lg px-2.5 py-1.5 text-xs transition-colors cursor-pointer",
-              sortOrder === opt.id
-                ? "bg-accent text-accent-foreground font-semibold"
-                : "text-popover-foreground hover:bg-accent/60 hover:text-accent-foreground font-medium",
-            )}
-          >
-            <span>{opt.label}</span>
-            {sortOrder === opt.id && (
-              <Check className="size-3.5 text-primary" />
-            )}
-          </button>
-        ))}
-      </div>
-    </div>
-  )
-}
-
-function ThreadListItem({
-  thread,
-  isActive,
-  isMenuOpen,
-  onSelect,
-  onStartRename,
-  onToggleArchive,
-  onDelete,
-  onToggleMenu,
-}: {
-  thread: ChatThreadPublic
-  isActive: boolean
-  isMenuOpen: boolean
-  onSelect: () => void
-  onStartRename: () => void
-  onToggleArchive: () => void
-  onDelete: () => void
-  onToggleMenu: () => void
-}) {
-  const isArchived = Boolean(thread.is_archived)
-
-  return (
-    <div
-      data-slot="thread-item"
-      className={cn(
-        "group relative flex items-center justify-between rounded-xl px-3 py-1.5 text-xs transition-colors cursor-pointer",
-        isActive
-          ? isArchived
-            ? "bg-muted/30 text-foreground/80 font-normal"
-            : "bg-muted/50 text-foreground font-medium"
-          : isArchived
-            ? "text-muted-foreground/70 hover:bg-muted/20 hover:text-muted-foreground"
-            : "text-muted-foreground hover:bg-muted/30 hover:text-foreground",
-      )}
-    >
-      <button
-        type="button"
-        onClick={onSelect}
-        className="flex flex-1 items-center truncate text-left cursor-pointer min-w-0 pr-2 focus:outline-none py-1 select-none"
-      >
-        <span
-          className={cn(
-            "truncate w-full",
-            isArchived
-              ? "text-xs text-muted-foreground font-normal group-hover:text-foreground"
-              : "text-xs text-foreground font-medium",
-          )}
-        >
-          {thread.title}
-        </span>
-      </button>
-
-      <div className="flex items-center gap-0.5 shrink-0">
-        {isArchived ? (
-          <Button
-            variant="ghost"
-            size="icon"
-            aria-label="Delete thread"
-            className="size-6 opacity-0 group-hover:opacity-100 hover:bg-destructive/10 rounded-full shrink-0 transition-opacity text-muted-foreground hover:text-destructive cursor-pointer"
-            onClick={(e) => {
-              e.stopPropagation()
-              onDelete()
-            }}
-          >
-            <Trash2 className="size-3" />
-          </Button>
-        ) : (
-          <Button
-            variant="ghost"
-            size="icon"
-            aria-label="Archive thread"
-            className="size-6 opacity-0 group-hover:opacity-100 hover:bg-muted/60 rounded-full shrink-0 transition-opacity text-muted-foreground hover:text-foreground cursor-pointer"
-            onClick={(e) => {
-              e.stopPropagation()
-              onToggleArchive()
-            }}
-          >
-            <Archive className="size-3" />
-          </Button>
-        )}
-
-        <div className="relative">
-          <Button
-            variant="ghost"
-            size="icon"
-            aria-label="Thread options"
-            className={cn(
-              "size-6 hover:bg-muted/60 rounded-full shrink-0 transition-opacity cursor-pointer text-muted-foreground hover:text-foreground",
-              isMenuOpen
-                ? "opacity-100 bg-muted/60 text-foreground"
-                : "opacity-0 group-hover:opacity-100",
-            )}
-            onClick={(e) => {
-              e.stopPropagation()
-              onToggleMenu()
-            }}
-          >
-            <MoreHorizontal className="size-3" />
-          </Button>
-
-          {isMenuOpen && (
-            <ThreadActionsMenu
-              isArchived={isArchived}
-              onToggleArchive={onToggleArchive}
-              onStartRename={onStartRename}
-              onDelete={onDelete}
-            />
-          )}
-        </div>
-      </div>
-    </div>
-  )
+function filterAndSortThreads(
+  threadList: ChatThreadPublic[],
+  searchQuery: string,
+  sortOrder: SortOption,
+): ChatThreadPublic[] {
+  let result = [...threadList]
+  if (searchQuery.trim()) {
+    const q = searchQuery.toLowerCase().trim()
+    result = result.filter((t) => t.title.toLowerCase().includes(q))
+  }
+  result.sort((a, b) => {
+    const timeA = a.created_at ? new Date(a.created_at).getTime() : 0
+    const timeB = b.created_at ? new Date(b.created_at).getTime() : 0
+    if (sortOrder === "recent") {
+      return timeB - timeA
+    }
+    if (sortOrder === "oldest") {
+      return timeA - timeB
+    }
+    if (sortOrder === "title") {
+      return a.title.localeCompare(b.title)
+    }
+    if (sortOrder === "messages") {
+      return (b.message_count ?? 0) - (a.message_count ?? 0)
+    }
+    return 0
+  })
+  return result
 }
 
 function AIPage() {
@@ -436,37 +91,27 @@ function AIPage() {
   )
   const [threadToRename, setThreadToRename] =
     React.useState<ChatThreadPublic | null>(null)
+  const [threadToDelete, setThreadToDelete] =
+    React.useState<ChatThreadPublic | null>(null)
   const [openMenuThreadId, setOpenMenuThreadId] = React.useState<string | null>(
     null,
   )
   const [recentsOpen, setRecentsOpen] = React.useState(true)
   const [archivedOpen, setArchivedOpen] = React.useState(true)
+  const [isSearchOpen, setIsSearchOpen] = React.useState(false)
+  const [searchQuery, setSearchQuery] = React.useState("")
+  const [sortOrder, setSortOrder] = React.useState<SortOption>("recent")
+  const [isSortMenuOpen, setIsSortMenuOpen] = React.useState(false)
 
   const promptInputRef = React.useRef<HTMLTextAreaElement>(null)
 
-  // Fetch available AI models from backend/proxy
   const { data: modelsData } = useQuery({
     queryKey: ["ai-models"],
     queryFn: () => AiThreadsService.listAiModels(),
   })
 
-  const AI_MODEL_STORAGE_KEY = "linkx_ai_selected_model"
-
   const [selectedModelId, setSelectedModelIdState] = React.useState<string>(
-    () => {
-      if (
-        typeof window !== "undefined" &&
-        typeof localStorage !== "undefined"
-      ) {
-        try {
-          const saved = localStorage.getItem(AI_MODEL_STORAGE_KEY)
-          if (saved) return saved
-        } catch {
-          // ignore
-        }
-      }
-      return "gemini-3.6-flash-high"
-    },
+    getInitialStoredModel,
   )
 
   const setSelectedModelId = React.useCallback((modelId: string) => {
@@ -480,7 +125,6 @@ function AIPage() {
     }
   }, [])
 
-  // Sync default model from backend if user has not explicitly chosen one
   React.useEffect(() => {
     if (!modelsData) return
     const saved =
@@ -490,9 +134,7 @@ function AIPage() {
 
     if (!saved) {
       const defaultId = modelsData.default_model || modelsData.data?.[0]?.id
-      if (defaultId) {
-        setSelectedModelId(defaultId)
-      }
+      if (defaultId) setSelectedModelId(defaultId)
     } else if (modelsData.data && modelsData.data.length > 0) {
       const exists = modelsData.data.some((m) => m.id === saved)
       if (!exists) {
@@ -502,26 +144,18 @@ function AIPage() {
     }
   }, [modelsData, setSelectedModelId])
 
-  // Confirmation dialog state for thread deletion
-  const [threadToDelete, setThreadToDelete] =
-    React.useState<ChatThreadPublic | null>(null)
-
-  // Local optimistic messages state during streaming
   const [localMessages, setLocalMessages] = React.useState<ChatUIMessage[]>([])
   const [pendingQuestion, setPendingQuestion] =
     React.useState<AskUserToolPart | null>(null)
 
-  // 1. Fetch user threads from PostgreSQL backend
   const { data: threadsData, isLoading: isThreadsLoading } = useQuery({
     queryKey: ["ai-threads"],
     queryFn: () => AiThreadsService.listChatThreads({ skip: 0, limit: 100 }),
   })
 
   const threads = threadsData?.data ?? []
-
   const initialLoadedRef = React.useRef(false)
 
-  // Auto-select first thread ONLY on initial mount when threads are first loaded
   React.useEffect(() => {
     if (threads.length > 0 && !initialLoadedRef.current) {
       setActiveThreadId(threads[0].id)
@@ -529,17 +163,15 @@ function AIPage() {
     }
   }, [threads])
 
-  // 2. Fetch active thread detail with full JSONB transcript
   const { data: activeThreadDetail } = useQuery({
     queryKey: ["ai-thread", activeThreadId],
     queryFn: () => AiThreadsService.getChatThread({ id: activeThreadId! }),
     enabled: !!activeThreadId,
   })
 
-  // Synchronize local messages from persistent database transcript when not streaming
   React.useEffect(() => {
     if (!isStreaming) {
-      if (activeThreadDetail?.transcript) {
+      if (activeThreadId && activeThreadDetail?.transcript) {
         const msgs =
           (activeThreadDetail.transcript as { messages?: ChatUIMessage[] })
             ?.messages || []
@@ -550,7 +182,6 @@ function AIPage() {
     }
   }, [activeThreadDetail, isStreaming, activeThreadId])
 
-  // 3. Mutations for thread persistence
   const createThreadMutation = useMutation({
     mutationFn: (prompt?: string) =>
       AiThreadsService.createChatThread({
@@ -595,148 +226,123 @@ function AIPage() {
     },
   })
 
-  const [isSortMenuOpen, setIsSortMenuOpen] = React.useState(false)
-
-  // Close menus on document click
   React.useEffect(() => {
-    const handleDocumentClick = () => {
+    function handleClickOutside() {
       setOpenMenuThreadId(null)
       setIsSortMenuOpen(false)
     }
     if (openMenuThreadId || isSortMenuOpen) {
-      document.addEventListener("click", handleDocumentClick)
+      document.addEventListener("click", handleClickOutside)
+      return () => document.removeEventListener("click", handleClickOutside)
     }
-    return () => document.removeEventListener("click", handleDocumentClick)
   }, [openMenuThreadId, isSortMenuOpen])
 
-  const handleNewChat = () => {
-    initialLoadedRef.current = true
-    setActiveThreadId(null)
-    setLocalMessages([])
-    setRecentsOpen(true)
-    setTimeout(() => {
-      promptInputRef.current?.focus()
-    }, 50)
-  }
+  function handleSendMessage(text: string) {
+    if (!text.trim() || isStreaming) return
 
-  const handleSendMessage = async (text: string) => {
-    const trimmed = text.trim()
-    if (!trimmed) return
-
-    let threadId = activeThreadId
-
-    // If no thread exists yet, create it in DB first
-    if (!threadId) {
-      const newThread = await createThreadMutation.mutateAsync(trimmed)
-      threadId = newThread.id
-    }
-
-    const userMessage: ChatUIMessage = {
-      id: `usr-${Date.now()}`,
+    const userMsg: ChatUIMessage = {
+      id: `local-user-${Date.now()}`,
       role: "user",
-      parts: [{ type: "text", text: trimmed }],
+      parts: [{ type: "text", text: text.trim() }],
+      createdAt: new Date().toISOString(),
     }
 
-    const assistantPlaceholder: ChatUIMessage = {
-      id: `asst-${Date.now()}`,
+    const assistantMsgId = `local-assistant-${Date.now()}`
+    const assistantMsg: ChatUIMessage = {
+      id: assistantMsgId,
       role: "assistant",
-      parts: [{ type: "thought", content: "" }],
+      parts: [],
+      createdAt: new Date().toISOString(),
     }
 
-    setLocalMessages((prev) => [...prev, userMessage, assistantPlaceholder])
+    setLocalMessages((prev) => [...prev, userMsg, assistantMsg])
+    setPendingQuestion(null)
 
-    startStream(
-      threadId,
-      trimmed,
-      {
-        onThought: (content) => {
-          setLocalMessages((prev) =>
-            prev.map((msg, idx) => {
-              if (idx !== prev.length - 1) return msg
-              const existingThought = msg.parts.find(
-                (p) => p.type === "thought",
-              )
-              const otherParts = msg.parts.filter((p) => p.type !== "thought")
-              const newContent =
-                (existingThought && "content" in existingThought
-                  ? existingThought.content
-                  : "") + content
-              return {
-                ...msg,
-                parts: [
-                  { type: "thought", content: newContent },
-                  ...otherParts,
-                ],
-              }
-            }),
-          )
-        },
-        onTextDelta: (content) => {
-          setLocalMessages((prev) =>
-            prev.map((msg, idx) => {
-              if (idx !== prev.length - 1) return msg
-              const existingTextPart = msg.parts.find((p) => p.type === "text")
-              const otherParts = msg.parts.filter((p) => p.type !== "text")
-              const newText =
-                (existingTextPart ? existingTextPart.text : "") + content
-              return {
-                ...msg,
-                parts: [...otherParts, { type: "text", text: newText }],
-              }
-            }),
-          )
-        },
-        onDraftArtifact: (artifact) => {
-          setLocalMessages((prev) =>
-            prev.map((msg, idx) =>
-              idx === prev.length - 1
-                ? {
-                    ...msg,
-                    parts: [...msg.parts, { type: "draft_artifact", artifact }],
-                  }
-                : msg,
-            ),
-          )
-        },
-        onError: (errorMsg) => {
-          setLocalMessages((prev) =>
-            prev.map((msg, idx) =>
-              idx === prev.length - 1
-                ? {
-                    ...msg,
-                    parts: [
-                      ...msg.parts,
-                      {
-                        type: "text",
-                        text: `⚠️ **Error generating response**: ${errorMsg}`,
-                      },
-                    ],
-                  }
-                : msg,
-            ),
-          )
-        },
-        onDone: () => {
-          queryClient.invalidateQueries({ queryKey: ["ai-thread", threadId] })
-          queryClient.invalidateQueries({ queryKey: ["ai-threads"] })
-        },
-      },
-      selectedModelId,
-    )
-  }
+    async function execute() {
+      let targetThreadId = activeThreadId
+      if (!targetThreadId) {
+        try {
+          const newThread = await createThreadMutation.mutateAsync(text.trim())
+          targetThreadId = newThread.id
+          setActiveThreadId(newThread.id)
+        } catch {
+          return
+        }
+      }
 
-  const handleConfirmRename = (newTitle: string) => {
-    if (threadToRename) {
-      updateThreadMutation.mutate(
-        { id: threadToRename.id, title: newTitle },
+      startStream(
+        targetThreadId,
+        text.trim(),
         {
-          onSuccess: () => setThreadToRename(null),
+          onThought: (content) => {
+            setLocalMessages((prev) =>
+              prev.map((msg) => {
+                if (msg.id !== assistantMsgId) return msg
+                const parts = [...msg.parts]
+                const lastPart = parts[parts.length - 1]
+                if (lastPart && lastPart.type === "thought") {
+                  lastPart.content += content
+                } else {
+                  parts.push({ type: "thought", content })
+                }
+                return { ...msg, parts }
+              }),
+            )
+          },
+          onTextDelta: (delta) => {
+            setLocalMessages((prev) =>
+              prev.map((msg) => {
+                if (msg.id !== assistantMsgId) return msg
+                const parts = [...msg.parts]
+                const lastPart = parts[parts.length - 1]
+                if (lastPart && lastPart.type === "text") {
+                  lastPart.text += delta
+                } else {
+                  parts.push({ type: "text", text: delta })
+                }
+                return { ...msg, parts }
+              }),
+            )
+          },
+          onDone: () => {
+            queryClient.invalidateQueries({ queryKey: ["ai-threads"] })
+            queryClient.invalidateQueries({
+              queryKey: ["ai-thread", targetThreadId],
+            })
+          },
         },
+        selectedModelId,
       )
     }
+
+    execute()
   }
 
-  const handleToggleArchive = (thread: ChatThreadPublic) => {
+  function handleQuestionAnswer(_toolCallId: string, answers: AskUserAnswer[]) {
+    const text = answers
+      .map((a) => a.answer)
+      .filter(Boolean)
+      .join("; ")
+    if (text) {
+      handleSendMessage(text)
+    }
+  }
+
+  function handleNewChat() {
+    stopStream()
+    setActiveThreadId(null)
+    setLocalMessages([])
+    setPendingQuestion(null)
+    setTimeout(() => promptInputRef.current?.focus(), 50)
+  }
+
+  function handleConfirmRename(title: string) {
+    if (!threadToRename) return
+    updateThreadMutation.mutate({ id: threadToRename.id, title })
+    setThreadToRename(null)
+  }
+
+  function handleToggleArchive(thread: ChatThreadPublic) {
     setOpenMenuThreadId(null)
     updateThreadMutation.mutate({
       id: thread.id,
@@ -744,73 +350,28 @@ function AIPage() {
     })
   }
 
-  const handleOpenDeleteDialog = (thread: ChatThreadPublic) => {
-    setOpenMenuThreadId(null)
-    setThreadToDelete(thread)
-  }
-
-  const handleConfirmDelete = () => {
-    if (threadToDelete) {
-      deleteThreadMutation.mutate(threadToDelete.id)
-    }
-  }
-
-  const handleQuestionAnswer = (
-    _toolCallId: string,
-    answers: AskUserAnswer[],
-  ) => {
-    const formattedAnswers = answers
-      .filter((a) => a.answer.trim().length > 0)
-      .map((a) => `• **${a.question}**\n  👉 ${a.answer}`)
-      .join("\n\n")
-
-    setPendingQuestion(null)
-    handleSendMessage(`Here are my preferences:\n\n${formattedAnswers}`)
-  }
-
-  const [sortOrder, setSortOrder] = React.useState<
-    "recent" | "oldest" | "title" | "messages"
-  >("recent")
-  const [searchQuery, setSearchQuery] = React.useState("")
-  const [isSearchOpen, setIsSearchOpen] = React.useState(false)
-
-  const processThreads = React.useCallback(
-    (threadList: ChatThreadPublic[]) => {
-      let result = [...threadList]
-      if (searchQuery.trim()) {
-        const query = searchQuery.toLowerCase().trim()
-        result = result.filter((t) => t.title.toLowerCase().includes(query))
-      }
-      result.sort((a, b) => {
-        if (sortOrder === "recent") {
-          const aTime = new Date(a.updated_at || a.created_at || 0).getTime()
-          const bTime = new Date(b.updated_at || b.created_at || 0).getTime()
-          return bTime - aTime
-        }
-        if (sortOrder === "oldest") {
-          const aTime = new Date(a.created_at || 0).getTime()
-          const bTime = new Date(b.created_at || 0).getTime()
-          return aTime - bTime
-        }
-        if (sortOrder === "title") {
-          return a.title.localeCompare(b.title)
-        }
-        if (sortOrder === "messages") {
-          return (b.message_count ?? 0) - (a.message_count ?? 0)
-        }
-        return 0
-      })
-      return result
-    },
-    [searchQuery, sortOrder],
+  const recentThreads = React.useMemo(
+    () =>
+      filterAndSortThreads(
+        threads.filter((t) => !t.is_archived),
+        searchQuery,
+        sortOrder,
+      ),
+    [threads, searchQuery, sortOrder],
   )
 
-  const recentThreads = processThreads(threads.filter((t) => !t.is_archived))
-  const archivedThreads = processThreads(threads.filter((t) => t.is_archived))
+  const archivedThreads = React.useMemo(
+    () =>
+      filterAndSortThreads(
+        threads.filter((t) => t.is_archived),
+        searchQuery,
+        sortOrder,
+      ),
+    [threads, searchQuery, sortOrder],
+  )
 
   return (
     <div className="flex w-full min-h-[calc(100vh-3.5rem)] lg:min-h-screen bg-background text-foreground">
-      {/* Rename Thread Popup Dialog */}
       <RenameThreadDialog
         thread={threadToRename}
         isOpen={Boolean(threadToRename)}
@@ -819,13 +380,16 @@ function AIPage() {
         onConfirm={handleConfirmRename}
       />
 
-      {/* Delete Thread Confirmation Dialog */}
       <DeleteThreadConfirmDialog
         thread={threadToDelete}
         isOpen={Boolean(threadToDelete)}
         isPending={deleteThreadMutation.isPending}
         onClose={() => setThreadToDelete(null)}
-        onConfirm={handleConfirmDelete}
+        onConfirm={() => {
+          if (threadToDelete) {
+            deleteThreadMutation.mutate(threadToDelete.id)
+          }
+        }}
       />
 
       {/* 1. Center Column: Active Chat Feed */}
@@ -878,7 +442,6 @@ function AIPage() {
           </MessageScrollerProvider>
         )}
 
-        {/* Docked Prompt Form at bottom */}
         <div className="mx-auto flex w-full max-w-2xl flex-col gap-2 px-4 pb-4 shrink-0">
           <PromptForm
             inputRef={promptInputRef}
@@ -894,204 +457,47 @@ function AIPage() {
       </div>
 
       {/* 2. Right Column: History Sidebar */}
-      <div className="hidden w-80 md:block shrink-0">
-        <div className="sticky top-0 self-start p-4 flex flex-col gap-4">
-          {/* Recents Section */}
-          <div className="flex flex-col gap-1">
-            <div className="flex items-center justify-between px-2 py-1">
-              <button
-                type="button"
-                onClick={() => setRecentsOpen((prev) => !prev)}
-                className="flex items-center gap-1.5 text-sm font-semibold text-foreground hover:text-muted-foreground transition-colors cursor-pointer select-none"
-              >
-                <span>Recents</span>
-                <ChevronDown
-                  className={cn(
-                    "size-3.5 text-muted-foreground transition-transform duration-200",
-                    !recentsOpen && "-rotate-90",
-                  )}
-                />
-              </button>
-
-              <div className="flex items-center gap-0.5">
-                <Button
-                  variant="ghost"
-                  size="icon"
-                  aria-label="Filter chats"
-                  onClick={() => setIsSearchOpen((prev) => !prev)}
-                  className={cn(
-                    "size-7 rounded-full cursor-pointer transition-colors",
-                    isSearchOpen || searchQuery
-                      ? "bg-muted/80 text-foreground"
-                      : "text-muted-foreground hover:text-foreground",
-                  )}
-                >
-                  <Search className="size-3.5" />
-                </Button>
-
-                <div className="relative">
-                  <Button
-                    variant="ghost"
-                    size="icon"
-                    aria-label="Sort options"
-                    onClick={(e) => {
-                      e.stopPropagation()
-                      setIsSortMenuOpen((prev) => !prev)
-                    }}
-                    className={cn(
-                      "size-7 rounded-full cursor-pointer transition-colors",
-                      isSortMenuOpen
-                        ? "bg-muted/80 text-foreground"
-                        : "text-muted-foreground hover:text-foreground",
-                    )}
-                  >
-                    <MoreHorizontal className="size-3.5" />
-                  </Button>
-                  {isSortMenuOpen && (
-                    <SortFilterMenu
-                      sortOrder={sortOrder}
-                      onSelectSort={(newOrder) => {
-                        setSortOrder(newOrder)
-                        setIsSortMenuOpen(false)
-                      }}
-                    />
-                  )}
-                </div>
-
-                <Button
-                  variant="ghost"
-                  size="icon"
-                  aria-label="New chat"
-                  onClick={handleNewChat}
-                  className="size-7 text-muted-foreground hover:text-foreground rounded-full cursor-pointer"
-                >
-                  <SquarePen className="size-3.5" />
-                </Button>
-              </div>
-            </div>
-
-            {(isSearchOpen || searchQuery) && (
-              <div className="px-1 py-1 animate-in fade-in-0 duration-150">
-                <div className="relative flex items-center">
-                  <Search className="absolute left-2.5 size-3 text-muted-foreground pointer-events-none" />
-                  <input
-                    type="text"
-                    value={searchQuery}
-                    onChange={(e) => setSearchQuery(e.target.value)}
-                    placeholder="Filter chats…"
-                    className="w-full rounded-lg bg-muted/30 border border-border/60 pl-7 pr-7 py-1 text-xs text-foreground placeholder:text-muted-foreground focus:outline-none focus:ring-1 focus:ring-ring"
-                  />
-                  {searchQuery && (
-                    <button
-                      type="button"
-                      aria-label="Clear filter"
-                      onClick={() => setSearchQuery("")}
-                      className="absolute right-2 text-muted-foreground hover:text-foreground cursor-pointer"
-                    >
-                      <X className="size-3" />
-                    </button>
-                  )}
-                </div>
-              </div>
-            )}
-
-            {recentsOpen && (
-              <div className="space-y-0.5">
-                {isThreadsLoading ? (
-                  <p className="px-3 py-1.5 text-xs text-muted-foreground animate-pulse">
-                    Loading chats…
-                  </p>
-                ) : recentThreads.length === 0 ? (
-                  <p className="px-3 py-1.5 text-xs text-muted-foreground">
-                    No recent chats
-                  </p>
-                ) : (
-                  recentThreads.map((thread) => (
-                    <ThreadListItem
-                      key={thread.id}
-                      thread={thread}
-                      isActive={thread.id === activeThreadId}
-                      isMenuOpen={openMenuThreadId === thread.id}
-                      onSelect={() => {
-                        if (thread.id !== activeThreadId) {
-                          stopStream()
-                          setActiveThreadId(thread.id)
-                        }
-                      }}
-                      onStartRename={() => {
-                        setOpenMenuThreadId(null)
-                        setThreadToRename(thread)
-                      }}
-                      onToggleArchive={() => handleToggleArchive(thread)}
-                      onDelete={() => handleOpenDeleteDialog(thread)}
-                      onToggleMenu={() =>
-                        setOpenMenuThreadId((prev) =>
-                          prev === thread.id ? null : thread.id,
-                        )
-                      }
-                    />
-                  ))
-                )}
-              </div>
-            )}
-          </div>
-
-          {/* Archived Section */}
-          <div className="flex flex-col gap-1 border-t border-border/40 pt-2.5">
-            <div className="flex items-center justify-between px-2 py-1">
-              <button
-                type="button"
-                onClick={() => setArchivedOpen((prev) => !prev)}
-                className="flex items-center gap-1.5 text-xs font-semibold text-muted-foreground hover:text-foreground transition-colors cursor-pointer select-none"
-              >
-                <span>Archived</span>
-                <ChevronDown
-                  className={cn(
-                    "size-3.5 text-muted-foreground transition-transform duration-200",
-                    !archivedOpen && "-rotate-90",
-                  )}
-                />
-              </button>
-            </div>
-
-            {archivedOpen && (
-              <div className="space-y-0.5">
-                {archivedThreads.length === 0 ? (
-                  <p className="px-3 py-1.5 text-xs text-muted-foreground">
-                    No archived chats
-                  </p>
-                ) : (
-                  archivedThreads.map((thread) => (
-                    <ThreadListItem
-                      key={thread.id}
-                      thread={thread}
-                      isActive={thread.id === activeThreadId}
-                      isMenuOpen={openMenuThreadId === thread.id}
-                      onSelect={() => {
-                        if (thread.id !== activeThreadId) {
-                          stopStream()
-                          setActiveThreadId(thread.id)
-                        }
-                      }}
-                      onStartRename={() => {
-                        setOpenMenuThreadId(null)
-                        setThreadToRename(thread)
-                      }}
-                      onToggleArchive={() => handleToggleArchive(thread)}
-                      onDelete={() => handleOpenDeleteDialog(thread)}
-                      onToggleMenu={() =>
-                        setOpenMenuThreadId((prev) =>
-                          prev === thread.id ? null : thread.id,
-                        )
-                      }
-                    />
-                  ))
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
+      <AIThreadsSidebar
+        recentThreads={recentThreads}
+        archivedThreads={archivedThreads}
+        activeThreadId={activeThreadId}
+        openMenuThreadId={openMenuThreadId}
+        isLoading={isThreadsLoading}
+        sortOrder={sortOrder}
+        searchQuery={searchQuery}
+        isSearchOpen={isSearchOpen}
+        isSortMenuOpen={isSortMenuOpen}
+        recentsOpen={recentsOpen}
+        archivedOpen={archivedOpen}
+        onSelectThread={(threadId) => {
+          if (threadId !== activeThreadId) {
+            stopStream()
+            setActiveThreadId(threadId)
+          }
+        }}
+        onNewChat={handleNewChat}
+        onStartRename={(t) => {
+          setOpenMenuThreadId(null)
+          setThreadToRename(t)
+        }}
+        onToggleArchive={handleToggleArchive}
+        onDeleteThread={(t) => {
+          setOpenMenuThreadId(null)
+          setThreadToDelete(t)
+        }}
+        onToggleMenu={(id) =>
+          setOpenMenuThreadId((prev) => (prev === id ? null : id))
+        }
+        onToggleSearch={() => setIsSearchOpen((prev) => !prev)}
+        onSearchChange={setSearchQuery}
+        onToggleSortMenu={() => setIsSortMenuOpen((prev) => !prev)}
+        onSelectSortOrder={(newOrder) => {
+          setSortOrder(newOrder)
+          setIsSortMenuOpen(false)
+        }}
+        onToggleRecents={() => setRecentsOpen((prev) => !prev)}
+        onToggleArchived={() => setArchivedOpen((prev) => !prev)}
+      />
     </div>
   )
 }
